requirements.txt
Flask==1.1.1
gunicorn==20.0
whitenoise==5.0
psycopg2-binary==2.8.4
